requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
ccxt==4.3.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
